% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_inputs}
\alias{read_inputs}
\title{Read a sequence of Shiny Inputs}
\usage{
read_inputs(
  inputid = NULL,
  nullarg = NA,
  type = c("chr", "dbl", "lgl", "int"),
  input_obj = get("input", envir = shiny::getDefaultReactiveDomain())
)
}
\arguments{
\item{inputid}{a vector of inputids to read}

\item{nullarg}{the value to return if the input value is \code{NULL} i.e. missing}

\item{type}{one of \code{('chr','dbl','lgl','int')} - specifies type of atomic vector to return}

\item{input_obj}{the input object from the Shiny session (hopefully.)}
}
\value{
a vector of values
}
\description{
Read a sequence of Shiny Inputs
}
\examples{
read_inputs(inputid = c('select_1','select_2'), nullarg = NA, type = 'chr')
}
