% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbind_dt.R
\name{unbind_dt}
\alias{unbind_dt}
\title{Unbind Shiny Inputs (server function)
When generating inputs reactively, it's sometimes necessary to unbind old inputs so that you can use the updated inputs in Shiny.
`\code{"unbind_dt_js()"} needs to be called in the UI so that the JS code is available in the session,
and then \code{unbind_dt()} should be called as needed in the server component (usually as part of an observeEvent or eventReactive)}
\usage{
unbind_dt(dt_name, session = session)
}
\arguments{
\item{dt_name}{String representing the table's output ID (i.e. to unbind in output$table_alpha, use "table_alpha" as the parameter)}
}
\value{
a call to the JS that unbinds tables.
}
\description{
Unbind Shiny Inputs (server function)
When generating inputs reactively, it's sometimes necessary to unbind old inputs so that you can use the updated inputs in Shiny.
`\code{"unbind_dt_js()"} needs to be called in the UI so that the JS code is available in the session,
and then \code{unbind_dt()} should be called as needed in the server component (usually as part of an observeEvent or eventReactive)
}
\examples{
unbind_dt('player_table')

}
