% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbind_dt.R
\name{unbind_dt_js}
\alias{unbind_dt_js}
\title{Unbind Shiny Inputs (UI function)}
\usage{
unbind_dt_js()
}
\value{
Adds JS to HTML head
}
\description{
When generating inputs reactively, it's sometimes necessary to unbind old inputs so that you can use the updated inputs in Shiny.
`\code{"unbind_dt_js()"} needs to be called in the UI so that the JS code is available in the session,
and then \code{unbind_dt()} should be called as needed in the server component (usually as part of an observeEvent or eventReactive)
}
\examples{
unbind_dt_js()

}
