% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_inputs}
\alias{read_inputs}
\title{Read a sequence of Shiny Inputs}
\usage{
read_inputs(inputid, nullarg = NA, type = c("chr", "dbl", "lgl", "int"))
}
\arguments{
\item{inputid}{a vector of inputids to read}

\item{nullarg}{the value to return if the input value is \code{NULL} i.e. missing}

\item{type}{one of \code{('chr','dbl','lgl','int')} - specifies type of vector to return}
}
\value{
a vector of values
}
\description{
Read a sequence of Shiny Inputs
}
\examples{
read_inputs(c('select_1','select_2'),nullarg = NA,type = 'chr')

}
